Move FamousDefs fixture out into its own file
